added all

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  */
 int print_last_digit(int n)
 {
-	int a;
+	long int a;
 
 	if (n < 0)
 		n *= -1;
